Refactored styled-components to their own file / folder
