+++ client/styledComponents/styledPhoto.jsx
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+const PhotoContainer = styled.div`
+  width: 250px;
+  height: 250px;
+`;
+
+const Image = styled.img`
+  height: 100%;
+  width: 100%;
+  object-fit: cover;
+`;
+
+export { PhotoContainer, Image };
